pyproject.toml
[project]
name = "epoch-letter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
    "resend>=2.12.0",
    "python-dotenv>=1.0.1",
]
